Separate pre-existing and compatibility tests
Create a separation between the pre-existing tests (now named unit
tests) and the newly created compatibility tests. Running `npm test` now
has the same behaviour as before, and `npm run test:compat` can be used
to run the compatibility tests.

The compatibility tests are also run in the CI, where, on top of running
tests against all supported ESLint version, the compatibility tests are
also ran on every supported Node.js (defined in `package.json` at line
68) version. That is, on every supported Node.js version, the plugin is
tested against every supported ESLint version.
