+++ .github/workflows/check.yml
@@ -71,6 +71,33 @@ jobs:
         if: ${{ always() }}
         with:
           file: ./_reports/coverage/lcov.info
+  test-compat:
+    name: Compatibility tests
+    runs-on: ubuntu-latest
+    needs:
+      - test
+    strategy:
+      fail-fast: false
+      matrix:
+        node-version:
+          - 12
+          - 14
+          - 16
+          - 18
+    steps:
+      - name: Checkout repository
+        uses: actions/checkout@2541b1294d2704b0964813337f33b291d3f8596b
+      - name: Install Node.js
+        uses: actions/setup-node@2fddd8803e2f5c9604345a0b591c3020ee971a93
+        with:
+          cache: npm
+          node-version: ${{ matrix.node-version }}
+      - name: Install compatible npm version
+        run: npm install --global npm@8.1.2
+      - name: Install Node.js dependencies
+        run: npm ci
+      - name: Run compatibility tests
+        run: npm run test:compat
   test-mutation:
     name: Mutation tests
     runs-on: ubuntu-latest
