+++ tests/unit/no-top-level-variables.test.ts
@@ -1,7 +1,7 @@
 import * as parser from '@typescript-eslint/parser';
 import {RuleTester} from 'eslint';
-import {trimTestCases} from '../lib/helpers';
-import {noTopLevelVariables} from '../lib/rules/noTopLevelVariables';
+import {trimTestCases} from '../../lib/helpers';
+import {noTopLevelVariables} from '../../lib/rules/noTopLevelVariables';
 
 const valid: RuleTester.ValidTestCase[] = [
   {
